Bugfix: resolve circular dependency
