+++ backend/services/tool_configuration_service.py
@@ -38,7 +38,7 @@ from database.knowledge_db import get_knowledge_name_map_by_index_names
 from mcpadapt.smolagents_adapter import _sanitize_function_name
 from services.file_management_service import get_llm_model, validate_urls_access
 from services.vectordatabase_service import get_embedding_model_by_index_name, get_rerank_model
-from services.remote_mcp_service import create_httpx_client
+from utils.http_client_utils import create_httpx_client
 from database.client import minio_client
 from services.image_service import get_video_understanding_model, get_vlm_model
 from nexent.monitor import set_monitoring_context, set_monitoring_operation
